StartUp/requirements.txt
aiofiles==22.1.0
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
cryptography==39.0.1
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.1.6
django-admin-star==1.0.4
django-allauth==0.52.0
django-crispy-forms==2.0
django-filter==22.1
django-location-field==2.1.0
django-phonenumber-field==7.0.2
django-seed==0.3.1
djangorestframework==3.14.0
drf-yasg==1.21.5
environs==8.0.0
Faker==17.6.0
frozenlist==1.3.3
geographiclib==2.0
geopy==2.3.0
googletrans==3.1.0a0
gTTS==2.3.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
magic-filter==1.0.9
MarkupSafe==2.1.2
marshmallow==3.19.0
multidict==6.0.3
numpy==1.24.1
oauthlib==3.2.2
packaging==22.0
pandas==1.5.2
phonenumberslite==8.13.7
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
toposort==1.10
uritemplate==4.1.1
urllib3==1.26.13
yarl==1.8.2
